+++ demo/src/main/java/br/liveo/ndrawer/MainActivity.java
@@ -19,6 +19,7 @@ import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.util.SparseIntArray;
+import android.view.Menu;
 import android.view.View;
 import android.widget.Toast;
 
@@ -26,6 +27,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import br.liveo.interfaces.OnItemClickListener;
+import br.liveo.interfaces.OnPrepareOptionsMenuLiveo;
 import br.liveo.navigationliveo.NavigationLiveo;
 
 public class MainActivity extends NavigationLiveo implements OnItemClickListener {
@@ -54,10 +56,10 @@ public class MainActivity extends NavigationLiveo implements OnItemClickListener
         // icons list items
         List<Integer> mListIconItem = new ArrayList<>();
         mListIconItem.add(0, R.drawable.ic_inbox_black_24dp);
-        mListIconItem.add(1, R.drawable.ic_star_black_24dp); //Item no icon set 0
-        mListIconItem.add(2, R.drawable.ic_send_black_24dp); //Item no icon set 0
+        mListIconItem.add(1, R.drawable.ic_star_black_24dp); //Item no icon set NavigationLiveo.NO_ICON
+        mListIconItem.add(2, R.drawable.ic_send_black_24dp); //Item no icon set NavigationLiveo.NO_ICON
         mListIconItem.add(3, R.drawable.ic_drafts_black_24dp);
-        mListIconItem.add(4, 0); //When the item is a subHeader the value of the icon 0
+        mListIconItem.add(4, NavigationLiveo.NO_ICON); //When the item is a subHeader the value of the icon NavigationLiveo.NO_ICON
         mListIconItem.add(5, R.drawable.ic_delete_black_24dp);
         mListIconItem.add(6, R.drawable.ic_report_black_24dp);
 
@@ -79,13 +81,20 @@ public class MainActivity extends NavigationLiveo implements OnItemClickListener
 
                 //{optional} - List Customization "If you remove these methods and the list will take his white standard color"
                 //.selectorCheck(R.drawable.selector_check) //Inform the background of the selected item color
-                //.colorItemDefault(R.color.nliveo_gray) //Inform the standard color name, icon and counter
+                //.colorItemDefault(R.color.nliveo_blue_colorPrimary) //Inform the standard color name, icon and counter
                 //.colorItemSelected(R.color.nliveo_purple_colorPrimary) //State the name of the color, icon and meter when it is selected
                 //.backgroundList(R.color.nliveo_black_light) //Inform the list of background color
                 //.colorLineSeparator(R.color.nliveo_transparent) //Inform the color of the subheader line
 
                 .footerItem(R.string.settings, R.drawable.ic_settings_black_24dp)
+
+                //{optional} - Footer Customization
+                //.footerNameColor(R.color.nliveo_blue_colorPrimary)
+                //.footerIconColor(R.color.nliveo_blue_colorPrimary)
+                //.footerBackground(R.color.nliveo_white)
+
                 .setOnClickUser(onClickPhoto)
+                .setOnPrepareOptionsMenu(onPrepare)
                 .setOnClickFooter(onClickFooter)
                 .build();
     }
@@ -99,6 +108,12 @@ public class MainActivity extends NavigationLiveo implements OnItemClickListener
             mFragmentManager.beginTransaction().replace(R.id.container, mFragment).commit();
         }
     }
+
+    private OnPrepareOptionsMenuLiveo onPrepare = new OnPrepareOptionsMenuLiveo() {
+        @Override
+        public void onPrepareOptionsMenu(Menu menu, int position, boolean visible) {
+        }
+    };
 
     private View.OnClickListener onClickPhoto = new View.OnClickListener() {
         @Override